backtrack on sane structs.
can't use nginx hashes because they're not meant to be modified at
any time.

rb-trees are too much intrusive and add lots of uncertainties.
i'm sure they can be made to work, but they're not clear enough
for exploratory code.

i'll switch to linear scannng of arrays.  performance isn't relevant
for the PoC, and i'll try to keep them encapsulated.